Update usage freq code
